chore: Rename typedef folder name
