+++ src/Core/Typedef/AllChannel.ts
