+++ src/Core/Typedef/AllResponse.ts
