+++ src/Core/Typedef/HttpHeader.ts
